+++ cloud-functions/deploy-firestore-triggers.sh
