+++ cloud-functions/deploy.sh
