+++ cloud-functions/make-rgb-fits.py
