--- firestore-stats-updater/requirements.txt
@@ -1,2 +1,0 @@
-google-cloud-firestore
-python-dateutil
