--- get-fits-header/requirements.txt
@@ -1,2 +1,0 @@
-Flask
-google-cloud-storage
