--- get-observation-list/requirements.txt
@@ -1,5 +1,0 @@
-Flask
-google-cloud-firestore
-google-cloud-storage
-pandas
-pendulum
