--- get-recent-observations/requirements.txt
@@ -1,4 +1,0 @@
-Flask
-google-cloud-firestore
-google-cloud-storage
-pandas
