--- lookup-field/deploy.sh
@@ -1,10 +1,0 @@
-#!/bin/bash -e
-
-TOPIC=${1:-lookup-field}
-
-gcloud functions deploy \
-                 "${TOPIC}" \
-                 --entry-point entry_point \
-                 --runtime python37 \
-                 --no-allow-unauthenticated \
-                 --trigger-http
