--- raw-file-uploaded/requirements.txt
@@ -1,5 +1,0 @@
-google-cloud-firestore
-google-cloud-storage
-python-dateutil
-tqdm
-git+https://github.com/panoptes/panoptes-pipeline@develop#egg=panoptes-pipeline
